refactor: change name for all Animate to better align with class name. Change all Graph Renderer to a class with static method so they can be extended
