+++ packages/client/src/pages/Algorithm/heap.jsx
@@ -5,7 +5,7 @@ import { AnalyzeRuntime } from "./AlgComponent/analyzeRuntime.jsx";
 import { SaveInputToLocalStorage } from "./AlgComponent/saveInputToLocalStorage";
 import HeapConcreteStrategy from "../../algorithm-solver/heapsolver.js";
 import Common from "./Common/common";
-import TreeGraphRenderer from "./Common/Canimate.js";
+import TreeGraphRenderer from "./Common/treerenderer.js";
 import { TreeAnimationData, Node } from "./Common/animationdata.js"
 import ResultsTable from "./AlgComponent/tableCreater";
